* math.c (math_gamma): explicit cast to double.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@49969 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ math.c
@@ -875,7 +875,7 @@ math_gamma(VALUE obj, VALUE x)
     if (isinf(d) && signbit(d)) domain_error("gamma");
     if (d == floor(d)) {
 	if (d < 0.0) domain_error("gamma");
-	if (1.0 <= d && d <= NGAMMA_TABLE) {
+	if (1.0 <= d && d <= (double)NGAMMA_TABLE) {
 	    return DBL2NUM(fact_table[(int)d - 1]);
 	}
     }
